% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchFunc.R
\name{batchCall}
\alias{batchCall}
\title{Call exon CNV.}
\usage{
batchCall(
  probFile,
  mergedCovFile,
  outDir,
  sampleId,
  geneName = NULL,
  cutoff = list(prob = 1e-04, pool.count = 1, baseline = 50, lowdepth = 10)
)
}
\arguments{
\item{probFile}{Path of probability file.}

\item{mergedCovFile}{Path of merged coverage file.}

\item{outDir}{Path of directory to write to.}

\item{sampleId}{Sample ids to call CNV, seperated by comma(,).}

\item{geneName}{Gene symbols to call CNV, seperated by comma(,).}

\item{cutoff}{A list of cutoff parameters to filter exon CNV, 'prob' is
probability, 'pool.count' is the least counts of sample of which gene has
CNV. 'baseline' is the least baseline depth to call CNV. 'lowdepth' is
least depth in all samples to call CNV.}
}
\description{
Call exon CNV.
}
