% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchFunc.R
\name{batchDepth}
\alias{batchDepth}
\title{Calulate depth.}
\usage{
batchDepth(bamFiles, bedFile, outDir, thread, mapqFilter)
}
\arguments{
\item{bamFiles}{A character string of BAM file paths seperated by comma(,).}

\item{bedFile}{A character string of BED file path.}

\item{outDir}{Path of directory to write to.}

\item{thread}{Integer, number of thread.}

\item{mapqFilter}{A numeric value, least map quality of reads to calculate.}
}
\description{
Calulate depth.
}
