% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchFunc.R
\name{batchExonBed}
\alias{batchExonBed}
\title{Export exon bed file.}
\usage{
batchExonBed(annoFile, outPath, gene)
}
\arguments{
\item{annoFile}{Path of annotation file.}

\item{outPath}{Path to write to.}

\item{gene}{A character string of gene symbols seperated by comma(,) to
export.}
}
\description{
Export exon bed file.
}
