% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchFunc.R
\name{batchExportVCF}
\alias{batchExportVCF}
\title{Export VCF file.}
\usage{
batchExportVCF(geneCNVFile, outPath)
}
\arguments{
\item{geneCNVFile}{Path of gene CNV result file.}

\item{outPath}{Path to write to.}
}
\description{
Export VCF file.
}
