% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchFunc.R
\name{batchFilterTrans}
\alias{batchFilterTrans}
\title{Filter transcripts.}
\usage{
batchFilterTrans(annoFile, outPath, transIdxFile, bedFile)
}
\arguments{
\item{annoFile}{Path of annotation file.}

\item{outPath}{Path to write to.}

\item{transIdxFile}{Path of index file.}

\item{bedFile}{Path of bed file.}
}
\description{
Filter transcripts.
}
