% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchFunc.R
\name{batchFit}
\alias{batchFit}
\title{Fit model to data.}
\usage{
batchFit(mergedCovFile, metricsFile, outPath, lowDepth)
}
\arguments{
\item{mergedCovFile}{Path of merged coverage file.}

\item{metricsFile}{Path of metrics file.}

\item{outPath}{Path to write to.}

\item{lowDepth}{A numeric value, regions that avaerage depth less than this
value will replaced by NA.}
}
\description{
Fit model to data.
}
