% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchFunc.R
\name{batchMerge}
\alias{batchMerge}
\title{Merge coverage files.}
\usage{
batchMerge(covFiles, outPath)
}
\arguments{
\item{covFiles}{A character string of coverage file paths seperated by
comma(,).}

\item{outPath}{Path to write to.}
}
\description{
Merge coverage files.
}
