% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchFunc.R
\name{batchMergeDefault}
\alias{batchMergeDefault}
\title{Calculate metrics.}
\usage{
batchMergeDefault(
  covFiles,
  metricsFile,
  outPath,
  defaultDir,
  defaultCovFiles,
  defaultMetricsFile,
  genderNum
)
}
\arguments{
\item{covFiles}{A character string of coverage file paths seperated by
comma(,).}

\item{metricsFile}{Path of metrics file.}

\item{outPath}{Path to write to.}

\item{defaultDir}{Path of directory storing defaultCovFiles and
defaultMetricsFile.}

\item{defaultCovFiles}{A character string of default coverage file paths
seperated by comma(,).}

\item{defaultMetricsFile}{Path of default metrics file.}

\item{genderNum}{Integer, if counts of samples shares the same gender is less
than genderNum, it will use default data as baseline.}
}
\description{
Calculate metrics.
}
