% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchFunc.R
\name{batchMetrics}
\alias{batchMetrics}
\title{Calculate metrics.}
\usage{
batchMetrics(bamFiles, covFiles, outPath)
}
\arguments{
\item{bamFiles}{A character string of BAM file paths seperated by comma(,).}

\item{covFiles}{A character string of coverage file paths seperated by
comma(,).}

\item{outPath}{Path to write to.}
}
\description{
Calculate metrics.
}
