% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchFunc.R
\name{batchPlot}
\alias{batchPlot}
\title{Plot exon CNV result}
\usage{
batchPlot(geneCNVFile, mergedCovFile, metricsFile, gene, outPath, draw)
}
\arguments{
\item{geneCNVFile}{Path of gene CNV result file.}

\item{mergedCovFile}{Path of merged coverage file.}

\item{metricsFile}{Path of metrics file.}

\item{gene}{A character string of gene symbols to plot, seperated by comma(,).}

\item{outPath}{Path to write to.}

\item{draw}{Number of genes to plot.}
}
\description{
Plot exon CNV result
}
