% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callECNV.R
\name{calculateLog2ratio}
\alias{calculateLog2ratio}
\title{Compute the log2 ratio between sample depth and baseline.}
\usage{
calculateLog2ratio(x, baseline, badDepth)
}
\arguments{
\item{x}{A numeric vector of sample depth.}

\item{baseline}{A numeric vector contains baseline depth of region with the
same order in sample depth vector.}

\item{badDepth}{A numeric value; a baseline depth low than badDepth will be
set to \code{NA}.}
}
\value{
A numeric vector of log2 ratio.
}
\description{
Input vectors of sample depth and baseline to compute the log2
  ratio.
}
\author{
Zhan-Ni Chen
}
