% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callECNV.R
\name{callExonCNV}
\alias{callExonCNV}
\title{Call exon CNVs.}
\usage{
callExonCNV(probFile, mergedCovFile, path = NULL, sample.id = NULL,
  cutoff = list(prob = 1e-04, pool.count = 1, baseline = 50, lowdepth =
  10))
}
\arguments{
\item{probFile}{Path of probability file.}

\item{mergedCovFile}{Path of merged coverage file.}

\item{path}{Directory path to write to.}

\item{sample.id}{Sample ids to call CNV. By default it will run in all
samples.}

\item{cutoff}{A list of cutoff parameters to filter exon CNV, 'prob' is
probability, 'pool.count' is the least counts of sample of which gene has
CNV. 'baseline' is the least baseline depth to call CNV. 'lowdepth' is
least depth in all samples to call CNV.}
}
\description{
Call exon CNVs.
}
\author{
Zhan-Ni Chen
}
