% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{callGenderByCov}
\alias{callGenderByCov}
\title{Call gender from coverage file.}
\usage{
callGenderByCov(x)
}
\arguments{
\item{x}{A character string of coverage file (<fileName>.cov) path.}
}
\value{
A character string, \emph{Unknow}, \emph{Female} or \emph{Male}.
}
\description{
Call gender by read depth of chromosome X and Y. Compare the
  chi-squared values obtained to infer whether the male or female assumption
  fits read depth better.
}
\author{
Zhan-Ni Chen
}
