% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{depthOfRegion}
\alias{depthOfRegion}
\alias{depthOfRegionMultiCore}
\title{Calculate depth of coverage in a given region.}
\usage{
depthOfRegion(region, bamPath, mapq.filter = 30)

depthOfRegionMultiCore(
  region,
  bamPath,
  thread,
  batch,
  mapq.filter = 30,
  tmpDir = NULL
)
}
\arguments{
\item{region}{a GenomicRanges object.}

\item{bamPath}{a character string of the BAM file path.}

\item{mapq.filter}{a non-negative integer specifying the minimum mapping
quality to include. BAM records with mapping qualities less than mapqFilter
are discarded.}

\item{thread}{an integer providing the number of threads just for
\code{depthOfRegionMultiCore()}.}

\item{batch}{an integer giving how many GRanges are performed in a batch just
for \code{depthOfRegionMultiCore()}.}

\item{tmpDir}{a character string of the directory path to write to.}
}
\value{
A GenomicRanges object with a column named depth representing the
  sequencing depth of this region.
}
\description{
Input a BAM file and a GenomicRanges object of given region,
  calculate the total bases overlapping this region. Sequencing depth of
  coverage of region equals total bases divide by width of region.

  The input BAM file need a BAM index file (<filename>.bai). The index file
  can be generated by \emph{samtools} or other BAM/SAM tools after sorting.
  Function will stop if the BAM file lose an index file or its index file is
  older than its own.

  \code{depthOfRegionMultiCore()} can perform in multithreading with number
  of threads you input. Multithreading uses \code{parLapply()} function in
  \pkg{parallel} package and it can run on Linux but not Windows system. And
  the \emph{batch} option limits the number of GRanges in a batch to run in a
  thread.
}
\author{
Zhan-Ni Chen
####### Calculate depth of coverage in a given region #######
}
