% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{exonToBed}
\alias{exonToBed}
\title{Bed of exon positions.}
\usage{
exonToBed(file, path = NULL, gene = NULL)
}
\arguments{
\item{file}{A character string of the annotation file's path.}

\item{path}{A character string of the file path to write to.}

\item{gene}{A character string or vector of gene symbols to export.}
}
\description{
Extract exon informations from the annotation tabular file and
  output a BED file with 4 columns indicating chromosome, start and end
  position, exon id.
}
\author{
Zhan-Ni Chen
}
