% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{filterTrans}
\alias{filterTrans}
\title{Filtering transcripts.}
\usage{
filterTrans(file, path = NULL, bed = NULL, index = NULL)
}
\arguments{
\item{file}{A character string of the annotation file's path.}

\item{path}{A character string of the file path to write to.}

\item{bed}{A character string of the BED file's path.}

\item{index}{A character string of the Index file's path.}
}
\description{
Filtering annotation file. If input an index file, which has two
  columns indicating \strong{gene symbol} and \emph{transcript name} without
  a header line, it will return transcripts based on names in the index file.
  If more than one transcript share the same name, it will return the
  transcript which has the most exons. If input a BED file, i.e. a
  target-capture region provided by the kit, it will filter out regions out
  of BED file.
}
\author{
Zhan-Ni Chen
}
