% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getParameters}
\alias{getParameters}
\title{Get bam file parameters.}
\usage{
getParameters(bamFiles, covFiles, path = NULL)
}
\arguments{
\item{bamFiles}{Paths of BAM files.}

\item{covFiles}{Paths of coverage files.}

\item{path}{Path to write to.}
}
\description{
Get bam file parameters.
}
