% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{mergerCovFiles}
\alias{mergerCovFiles}
\title{Merge coverage files.}
\usage{
mergerCovFiles(x, path = NULL)
}
\arguments{
\item{x}{A character vector contains several coverage files path.}

\item{path}{Path to write to.}
}
\value{
A data frame, of which columns are chromosome, start position, end
  position, id and depths in input coverage files.
}
\description{
Align and merge coverage files (<filename>.cov) with chromosome,
  start and end position. Five required fields in a coverage file are
  chromosome name, start and end position, id and depth of coverage.
}
\author{
Zhan-Ni Chen
####### Merge coverage files #######
}
