% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCNV.R
\name{plotExonCNV}
\alias{plotExonCNV}
\title{Plot the unifomity of coverage among samples.}
\usage{
plotExonCNV(
  geneCNVFile,
  mergedCovFile,
  parameterFile,
  gene = NULL,
  path = NULL,
  draw = 10
)
}
\arguments{
\item{geneCNVFile}{Path of gene cnv result file.}

\item{mergedCovFile}{Path of merged coverage file.}

\item{parameterFile}{Path of metrics file.}

\item{gene}{A character string or vector of gene symbol.}

\item{path}{Path to write to.}

\item{draw}{An integer of gene counts to plot.}
}
\description{
Compute correlation between columns of a numeric matrix or data
  frame and plot the distribution of correlations.
}
\author{
Zhan-Ni Chen
}
