% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchFunc.R
\name{batchExonBed}
\alias{batchExonBed}
\title{Export exon bed file.}
\usage{
batchExonBed(annoFile, outPath, gene, expand)
}
\arguments{
\item{annoFile}{Path of annotation file.}

\item{outPath}{Path to write to.}

\item{gene}{A character string of gene symbols seperated by comma(,) to
export.}

\item{expand}{An integer. If an exon's size is less than this value, its
region will be expanded centrally to a width of this value.}
}
\description{
Export exon bed file.
}
